hw2-2 mpi test 30

+++ hw2-2/F74102022_hw2_2.c
@@ -10,7 +10,6 @@
 void custom_min(void *in, void *inout, int *len, MPI_Datatype *datatype) {
     int *in_array = (int *)in;
     int *inout_array = (int *)inout;
-//printf("index:%d dict:%d VS. index:%d dict:%d \n", inout_array[0], inout_array[1], in_array[0], in_array[1]);
     if (in_array[1] < inout_array[1]) {
         inout_array[1] = in_array[1];
         inout_array[0] = in_array[0];
@@ -22,7 +21,7 @@ int main( int argc, char *argv[]){
     char input[50];
     int size;
     bool selected[50000];
-    int dist[50000];   
+    short dist[50000];   
     int min[2];         //(index, dist)
     short **Adj;
